+++ channeld/channeld_wire.csv
@@ -224,23 +224,3 @@ msgdata,channeld_send_error,reason,wirestring,
 
 # Tell master channeld has sent the error message.
 msgtype,channeld_send_error_reply,1108
-
-# Tell lightningd we got a onion message (for us, or to fwd)
-msgtype,got_onionmsg_to_us,1142
-msgdata,got_onionmsg_to_us,blinding_in,?pubkey,
-msgdata,got_onionmsg_to_us,reply_blinding,?pubkey,
-msgdata,got_onionmsg_to_us,reply_path_len,u16,
-msgdata,got_onionmsg_to_us,reply_path,onionmsg_path,reply_path_len
-msgdata,got_onionmsg_to_us,rawmsg_len,u16,
-msgdata,got_onionmsg_to_us,rawmsg,u8,rawmsg_len
-
-msgtype,got_onionmsg_forward,1143
-msgdata,got_onionmsg_forward,next_scid,?short_channel_id,
-msgdata,got_onionmsg_forward,next_node_id,?node_id,
-msgdata,got_onionmsg_forward,next_blinding,?pubkey,
-msgdata,got_onionmsg_forward,next_onion,u8,1366
-
-# Lightningd tells us to send a onion message.
-msgtype,send_onionmsg,1040
-msgdata,send_onionmsg,onion,u8,1366
-msgdata,send_onionmsg,blinding,?pubkey,
