+++ gossipd/Makefile
@@ -33,6 +33,7 @@ GOSSIPD_COMMON_OBJS :=				\
 	common/bech32_util.o			\
 	common/bigsize.o			\
 	common/bip32.o				\
+	common/blinding.o			\
 	common/channel_id.o			\
 	common/crypto_state.o			\
 	common/cryptomsg.o			\
@@ -41,13 +42,16 @@ GOSSIPD_COMMON_OBJS :=				\
 	common/decode_array.o			\
 	common/derive_basepoints.o		\
 	common/dev_disconnect.o			\
+	common/ecdh_hsmd.o			\
 	common/features.o			\
+	common/hmac.o				\
 	common/status_wiregen.o			\
 	common/gossip_rcvd_filter.o		\
 	common/key_derive.o			\
 	common/memleak.o			\
 	common/msg_queue.o			\
 	common/node_id.o			\
+	common/onion.o				\
 	common/onionreply.o			\
 	common/per_peer_state.o			\
 	common/ping.o				\
@@ -56,6 +60,7 @@ GOSSIPD_COMMON_OBJS :=				\
 	common/private_channel_announcement.o	\
 	common/random_select.o			\
 	common/setup.o				\
+	common/sphinx.o				\
 	common/status.o				\
 	common/status_wire.o			\
 	common/subdaemon.o			\
