+++ gossipd/gossipd_wiregen.c
@@ -46,6 +46,9 @@ const char *gossipd_wire_name(int e)
 	case WIRE_GOSSIPD_GET_INCOMING_CHANNELS: return "WIRE_GOSSIPD_GET_INCOMING_CHANNELS";
 	case WIRE_GOSSIPD_GET_INCOMING_CHANNELS_REPLY: return "WIRE_GOSSIPD_GET_INCOMING_CHANNELS_REPLY";
 	case WIRE_GOSSIPD_NEW_BLOCKHEIGHT: return "WIRE_GOSSIPD_NEW_BLOCKHEIGHT";
+	case WIRE_GOSSIPD_GOT_ONIONMSG_TO_US: return "WIRE_GOSSIPD_GOT_ONIONMSG_TO_US";
+	case WIRE_GOSSIPD_GOT_ONIONMSG_FORWARD: return "WIRE_GOSSIPD_GOT_ONIONMSG_FORWARD";
+	case WIRE_GOSSIPD_SEND_ONIONMSG: return "WIRE_GOSSIPD_SEND_ONIONMSG";
 	}
 
 	snprintf(invalidbuf, sizeof(invalidbuf), "INVALID %i", e);
@@ -81,6 +84,9 @@ bool gossipd_wire_is_defined(u16 type)
 	case WIRE_GOSSIPD_GET_INCOMING_CHANNELS:;
 	case WIRE_GOSSIPD_GET_INCOMING_CHANNELS_REPLY:;
 	case WIRE_GOSSIPD_NEW_BLOCKHEIGHT:;
+	case WIRE_GOSSIPD_GOT_ONIONMSG_TO_US:;
+	case WIRE_GOSSIPD_GOT_ONIONMSG_FORWARD:;
+	case WIRE_GOSSIPD_SEND_ONIONMSG:;
 	      return true;
 	}
 	return false;
@@ -857,4 +863,160 @@ bool fromwire_gossipd_new_blockheight(const void *p, u32 *blockheight)
  	*blockheight = fromwire_u32(&cursor, &plen);
 	return cursor != NULL;
 }
-// SHA256STAMP:ae8bf4f19cc3dd086b9a20b8bc034d93d8ef066e901279de09fe2fd189f1344b
+
+/* WIRE: GOSSIPD_GOT_ONIONMSG_TO_US */
+/* Tell lightningd we got a onion message (for us */
+u8 *towire_gossipd_got_onionmsg_to_us(const tal_t *ctx, const struct pubkey *blinding_in, const struct pubkey *reply_blinding, const struct onionmsg_path **reply_path, const u8 *rawmsg)
+{
+	u16 reply_path_len = tal_count(reply_path);
+	u16 rawmsg_len = tal_count(rawmsg);
+	u8 *p = tal_arr(ctx, u8, 0);
+
+	towire_u16(&p, WIRE_GOSSIPD_GOT_ONIONMSG_TO_US);
+	if (!blinding_in)
+		towire_bool(&p, false);
+	else {
+		towire_bool(&p, true);
+		towire_pubkey(&p, blinding_in);
+	}
+	if (!reply_blinding)
+		towire_bool(&p, false);
+	else {
+		towire_bool(&p, true);
+		towire_pubkey(&p, reply_blinding);
+	}
+	towire_u16(&p, reply_path_len);
+	for (size_t i = 0; i < reply_path_len; i++)
+		towire_onionmsg_path(&p, reply_path[i]);
+	towire_u16(&p, rawmsg_len);
+	towire_u8_array(&p, rawmsg, rawmsg_len);
+
+	return memcheck(p, tal_count(p));
+}
+bool fromwire_gossipd_got_onionmsg_to_us(const tal_t *ctx, const void *p, struct pubkey **blinding_in, struct pubkey **reply_blinding, struct onionmsg_path ***reply_path, u8 **rawmsg)
+{
+	u16 reply_path_len;
+	u16 rawmsg_len;
+
+	const u8 *cursor = p;
+	size_t plen = tal_count(p);
+
+	if (fromwire_u16(&cursor, &plen) != WIRE_GOSSIPD_GOT_ONIONMSG_TO_US)
+		return false;
+ 	if (!fromwire_bool(&cursor, &plen))
+		*blinding_in = NULL;
+	else {
+		*blinding_in = tal(ctx, struct pubkey);
+		fromwire_pubkey(&cursor, &plen, *blinding_in);
+	}
+ 	if (!fromwire_bool(&cursor, &plen))
+		*reply_blinding = NULL;
+	else {
+		*reply_blinding = tal(ctx, struct pubkey);
+		fromwire_pubkey(&cursor, &plen, *reply_blinding);
+	}
+ 	reply_path_len = fromwire_u16(&cursor, &plen);
+ 	// 2nd case reply_path
+	*reply_path = reply_path_len ? tal_arr(ctx, struct onionmsg_path *, reply_path_len) : NULL;
+	for (size_t i = 0; i < reply_path_len; i++)
+		(*reply_path)[i] = fromwire_onionmsg_path(*reply_path, &cursor, &plen);
+ 	rawmsg_len = fromwire_u16(&cursor, &plen);
+ 	// 2nd case rawmsg
+	*rawmsg = rawmsg_len ? tal_arr(ctx, u8, rawmsg_len) : NULL;
+	fromwire_u8_array(&cursor, &plen, *rawmsg, rawmsg_len);
+	return cursor != NULL;
+}
+
+/* WIRE: GOSSIPD_GOT_ONIONMSG_FORWARD */
+u8 *towire_gossipd_got_onionmsg_forward(const tal_t *ctx, const struct short_channel_id *next_scid, const struct node_id *next_node_id, const struct pubkey *next_blinding, const u8 next_onion[1366])
+{
+	u8 *p = tal_arr(ctx, u8, 0);
+
+	towire_u16(&p, WIRE_GOSSIPD_GOT_ONIONMSG_FORWARD);
+	if (!next_scid)
+		towire_bool(&p, false);
+	else {
+		towire_bool(&p, true);
+		towire_short_channel_id(&p, next_scid);
+	}
+	if (!next_node_id)
+		towire_bool(&p, false);
+	else {
+		towire_bool(&p, true);
+		towire_node_id(&p, next_node_id);
+	}
+	if (!next_blinding)
+		towire_bool(&p, false);
+	else {
+		towire_bool(&p, true);
+		towire_pubkey(&p, next_blinding);
+	}
+	towire_u8_array(&p, next_onion, 1366);
+
+	return memcheck(p, tal_count(p));
+}
+bool fromwire_gossipd_got_onionmsg_forward(const tal_t *ctx, const void *p, struct short_channel_id **next_scid, struct node_id **next_node_id, struct pubkey **next_blinding, u8 next_onion[1366])
+{
+	const u8 *cursor = p;
+	size_t plen = tal_count(p);
+
+	if (fromwire_u16(&cursor, &plen) != WIRE_GOSSIPD_GOT_ONIONMSG_FORWARD)
+		return false;
+ 	if (!fromwire_bool(&cursor, &plen))
+		*next_scid = NULL;
+	else {
+		*next_scid = tal(ctx, struct short_channel_id);
+		fromwire_short_channel_id(&cursor, &plen, *next_scid);
+	}
+ 	if (!fromwire_bool(&cursor, &plen))
+		*next_node_id = NULL;
+	else {
+		*next_node_id = tal(ctx, struct node_id);
+		fromwire_node_id(&cursor, &plen, *next_node_id);
+	}
+ 	if (!fromwire_bool(&cursor, &plen))
+		*next_blinding = NULL;
+	else {
+		*next_blinding = tal(ctx, struct pubkey);
+		fromwire_pubkey(&cursor, &plen, *next_blinding);
+	}
+ 	fromwire_u8_array(&cursor, &plen, next_onion, 1366);
+	return cursor != NULL;
+}
+
+/* WIRE: GOSSIPD_SEND_ONIONMSG */
+/* Lightningd tells us to send a onion message. */
+u8 *towire_gossipd_send_onionmsg(const tal_t *ctx, const struct node_id *id, const u8 onion[1366], const struct pubkey *blinding)
+{
+	u8 *p = tal_arr(ctx, u8, 0);
+
+	towire_u16(&p, WIRE_GOSSIPD_SEND_ONIONMSG);
+	towire_node_id(&p, id);
+	towire_u8_array(&p, onion, 1366);
+	if (!blinding)
+		towire_bool(&p, false);
+	else {
+		towire_bool(&p, true);
+		towire_pubkey(&p, blinding);
+	}
+
+	return memcheck(p, tal_count(p));
+}
+bool fromwire_gossipd_send_onionmsg(const tal_t *ctx, const void *p, struct node_id *id, u8 onion[1366], struct pubkey **blinding)
+{
+	const u8 *cursor = p;
+	size_t plen = tal_count(p);
+
+	if (fromwire_u16(&cursor, &plen) != WIRE_GOSSIPD_SEND_ONIONMSG)
+		return false;
+ 	fromwire_node_id(&cursor, &plen, id);
+ 	fromwire_u8_array(&cursor, &plen, onion, 1366);
+ 	if (!fromwire_bool(&cursor, &plen))
+		*blinding = NULL;
+	else {
+		*blinding = tal(ctx, struct pubkey);
+		fromwire_pubkey(&cursor, &plen, *blinding);
+	}
+	return cursor != NULL;
+}
+// SHA256STAMP:45335ac4c553938ed7c2d63344d80465eca1ff70ab8ec750e3d0305f81acdad5
