+++ gossipd/gossipd_wiregen.h
@@ -60,6 +60,11 @@ enum gossipd_wire {
         WIRE_GOSSIPD_GET_INCOMING_CHANNELS_REPLY = 3125,
         /*  master -> gossipd: blockheight increased. */
         WIRE_GOSSIPD_NEW_BLOCKHEIGHT = 3026,
+        /*  Tell lightningd we got a onion message (for us */
+        WIRE_GOSSIPD_GOT_ONIONMSG_TO_US = 3142,
+        WIRE_GOSSIPD_GOT_ONIONMSG_FORWARD = 3143,
+        /*  Lightningd tells us to send a onion message. */
+        WIRE_GOSSIPD_SEND_ONIONMSG = 3040,
 };
 
 const char *gossipd_wire_name(int e);
@@ -197,6 +202,20 @@ bool fromwire_gossipd_get_incoming_channels_reply(const tal_t *ctx, const void *
 u8 *towire_gossipd_new_blockheight(const tal_t *ctx, u32 blockheight);
 bool fromwire_gossipd_new_blockheight(const void *p, u32 *blockheight);
 
+/* WIRE: GOSSIPD_GOT_ONIONMSG_TO_US */
+/*  Tell lightningd we got a onion message (for us */
+u8 *towire_gossipd_got_onionmsg_to_us(const tal_t *ctx, const struct pubkey *blinding_in, const struct pubkey *reply_blinding, const struct onionmsg_path **reply_path, const u8 *rawmsg);
+bool fromwire_gossipd_got_onionmsg_to_us(const tal_t *ctx, const void *p, struct pubkey **blinding_in, struct pubkey **reply_blinding, struct onionmsg_path ***reply_path, u8 **rawmsg);
+
+/* WIRE: GOSSIPD_GOT_ONIONMSG_FORWARD */
+u8 *towire_gossipd_got_onionmsg_forward(const tal_t *ctx, const struct short_channel_id *next_scid, const struct node_id *next_node_id, const struct pubkey *next_blinding, const u8 next_onion[1366]);
+bool fromwire_gossipd_got_onionmsg_forward(const tal_t *ctx, const void *p, struct short_channel_id **next_scid, struct node_id **next_node_id, struct pubkey **next_blinding, u8 next_onion[1366]);
+
+/* WIRE: GOSSIPD_SEND_ONIONMSG */
+/*  Lightningd tells us to send a onion message. */
+u8 *towire_gossipd_send_onionmsg(const tal_t *ctx, const struct node_id *id, const u8 onion[1366], const struct pubkey *blinding);
+bool fromwire_gossipd_send_onionmsg(const tal_t *ctx, const void *p, struct node_id *id, u8 onion[1366], struct pubkey **blinding);
+
 
 #endif /* LIGHTNING_GOSSIPD_GOSSIPD_WIREGEN_H */
-// SHA256STAMP:ae8bf4f19cc3dd086b9a20b8bc034d93d8ef066e901279de09fe2fd189f1344b
+// SHA256STAMP:45335ac4c553938ed7c2d63344d80465eca1ff70ab8ec750e3d0305f81acdad5
