+++ lightningd/onion_message.h
@@ -3,9 +3,9 @@
 #include "config.h"
 #include <ccan/short_types/short_types.h>
 
-struct channel;
+struct lightningd;
 
-void handle_onionmsg_to_us(struct channel *channel, const u8 *msg);
-void handle_onionmsg_forward(struct channel *channel, const u8 *msg);
+void handle_onionmsg_to_us(struct lightningd *ld, const u8 *msg);
+void handle_onionmsg_forward(struct lightningd *ld, const u8 *msg);
 
 #endif /* LIGHTNING_LIGHTNINGD_ONION_MESSAGE_H */
